Made a number of updates and improvements.
Updated and added logs. Updated the path according to RESTfull conventions. Removed redundant code and classes. Adapted the tests to the changes.

+++ src/main/java/ua/lastbite/email_service/exception/UserNotFoundException.java
@@ -1,7 +1,7 @@
 package ua.lastbite.email_service.exception;
 
 public class UserNotFoundException extends RuntimeException {
-    public UserNotFoundException(int id) {
+    public UserNotFoundException(long id) {
         super("User with ID " + id + " not found");
     }
 }
